Make sure the condition check not fail
- Wrap into the block with condition when ami_image is supplied.
- Adjust conditions when searching for ami using search tag

This is unknown to me yet why ansible return an error if it is called from role
ec2_instance and it reaches the code in the second point above. Even debugging
show the ami_search_tag is undefined some how when executing the module, it
goes back to the original role (ec2_instance) and tries to parse the default
value of ami_search_tag and got error during the process.

That means 'debug' has a different behaviour with 'when' - that is strange.